refactor(core): add to_openai_format methods to Message and ContentPart
Move OpenAI format conversion logic to the domain types themselves:
- Message::to_openai_format() - converts message to OpenAI-compatible JSON
- ContentPart::to_openai_format() - converts content part to OpenAI format

The standalone functions in observation/openai_format.rs now delegate
to these methods. This is more idiomatic Rust and consistent with
existing Message methods like content_to_llm_string().
